[feat] feed 삭제 api / DetailFeedViewModel -> MainViewModel로 통합(추후 이벤트 처리 위함) (#105)
