+++ feature/main/src/main/java/com/readme/android/main/viewmodel/MainViewModel.kt
@@ -9,6 +9,7 @@ import com.readme.android.core_ui.base.BaseViewModel
 import com.readme.android.core_ui.util.Event
 import com.readme.android.core_ui.util.MutableEventFlow
 import com.readme.android.core_ui.util.asEventFlow
+import com.readme.android.domain.entity.BookInfo
 import com.readme.android.domain.entity.FeedInfo
 import com.readme.android.domain.repository.FeedRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -81,6 +82,38 @@ class MainViewModel @Inject constructor(
 
     private fun categoryListToString(): String =
         selectedCategoryChip.value?.joinToString(SEPARATOR) ?: ""
+
+    /** FeedDetailActivity */
+    private val feedId = MutableLiveData<Int>()
+
+    private val _feed = MutableLiveData<FeedInfo>()
+    val feed: LiveData<FeedInfo> = _feed
+
+    private val _bookInfo = MutableLiveData<BookInfo>()
+    val bookInfo: LiveData<BookInfo> = _bookInfo
+
+    fun setFeedId(feedId: Int) {
+        this.feedId.value = feedId
+    }
+
+    fun getIsMyFeed(): Boolean = feed.value?.isMyFeed ?: throw IllegalStateException()
+
+    fun getWriterNickname(): String =
+        feed.value?.nickname ?: throw IllegalStateException("writer nickname cannot be null")
+
+    fun getFeedId(): Int =
+        feed.value?.id ?: throw IllegalStateException("feed id cannot be null")
+
+    fun getDetailFeedInfo() {
+        viewModelScope.launch(exceptionHandler) {
+            feedRepository.getDetailFeed(feedId.value ?: throw IllegalStateException())
+                .onSuccess {
+                    _feed.value = it.feed
+                    _bookInfo.value = it.bookInfo
+                }
+                .onFailure { Log.d(TAG, "getDetailFeedInfo: $it") }
+        }
+    }
 
     /** server event */
     private val _isNetworkCorrespondenceEnd = MutableLiveData<Event<String>>()
